remove UI flick
