+++ app/(auth)/transactions/loading.tsx
@@ -1,0 +1,5 @@
+import { DotLoader } from "@/components/ui/skeleton/dot-loader";
+
+export default function Loading() {
+  return <DotLoader />;
+}
