+++ components/ui/skeleton/dot-loader.tsx
@@ -1,9 +1,6 @@
 import React from 'react'
 
-export const DotLoader = ({ isShow }: { isShow: boolean }) => {
-    if (!isShow) {
-        return null
-    }
+export const DotLoader = () => {
     return (
         <div className="loader-dots absolute inset-1/2 -translate-x-1/2 -translate-y-1/2" />
     )
